feat: add initial support for esp32-s3
