+++ src/runtime/runtime_esp32s3.go
@@ -3,146 +3,57 @@
 package runtime
 
 import (
-	"device"
 	"device/esp"
-	"machine"
-	"unsafe"
 )
 
-type timeUnit int64
-
-func putchar(c byte) {
-	machine.Serial.WriteByte(c)
-}
-
-func getchar() byte {
-	for machine.Serial.Buffered() == 0 {
-		Gosched()
-	}
-	v, _ := machine.Serial.ReadByte()
-	return v
-}
-
-func buffered() int {
-	return machine.Serial.Buffered()
-}
-
-// Initialize .bss: zero-initialized global variables.
-// The .data section has already been loaded by the ROM bootloader.
-func clearbss() {
-	ptr := unsafe.Pointer(&_sbss)
-	for ptr != unsafe.Pointer(&_ebss) {
-		*(*uint32)(ptr) = 0
-		ptr = unsafe.Add(ptr, 4)
-	}
-}
-
-func initTimer() {
-	// Configure timer 0 in timer group 0, for timekeeping.
-	//   EN:       Enable the timer.
-	//   INCREASE: Count up every tick (as opposed to counting down).
-	//   DIVIDER:  16-bit prescaler, set to 2 for dividing the APB clock by two
-	//             (40MHz).
-	// esp.TIMG0.T0CONFIG.Set(0 << esp.TIMG_T0CONFIG_T0_EN_Pos)
-	// TODO: Add support for esp32s3
-	// esp.TIMG0.T0CONFIG.Set(esp.TIMG_T0CONFIG_T0_EN | esp.TIMG_T0CONFIG_T0_INCREASE | 2<<esp.TIMG_T0CONFIG_T0_DIVIDER_Pos)
-
-	// esp.TIMG0.T0CONFIG.Set(1 << esp.TIMG_T0CONFIG_T0_DIVCNT_RST_Pos)
-	// esp.TIMG0.T0CONFIG.Set(esp.TIMG_T0CONFIG_T0_EN)
-
-	// Set the timer counter value to 0.
-	esp.TIMG0.T0LOADLO.Set(0)
-	esp.TIMG0.T0LOADHI.Set(0)
-	esp.TIMG0.T0LOAD.Set(0) // value doesn't matter.
-}
-
-func ticks() timeUnit {
-	// First, update the LO and HI register pair by writing any value to the
-	// register. This allows reading the pair atomically.
-	esp.TIMG0.T0UPDATE.Set(0)
-	// Then read the two 32-bit parts of the timer.
-	return timeUnit(uint64(esp.TIMG0.T0LO.Get()) | uint64(esp.TIMG0.T0HI.Get())<<32)
-}
-
-func nanosecondsToTicks(ns int64) timeUnit {
-	// Calculate the number of ticks from the number of nanoseconds. At a 80MHz
-	// APB clock, that's 25 nanoseconds per tick with a timer prescaler of 2:
-	// 25 = 1e9 / (80MHz / 2)
-	return timeUnit(ns / 25)
-}
-
-func ticksToNanoseconds(ticks timeUnit) int64 {
-	// See nanosecondsToTicks.
-	return int64(ticks) * 25
-}
-
-// sleepTicks busy-waits until the given number of ticks have passed.
-func sleepTicks(d timeUnit) {
-	sleepUntil := ticks() + d
-	for ticks() < sleepUntil {
-		// TODO: suspend the CPU to not burn power here unnecessarily.
-	}
-}
-
-func exit(code int) {
-	abort()
-}
-
-//go:extern _sbss
-var _sbss [0]byte
-
-//go:extern _ebss
-var _ebss [0]byte
-
-func abort() {
-	for {
-		device.Asm("waiti 0")
-	}
-}
-
-// This is the function called on startup right after the stack pointer has been
-// set.
+// This is the function called on startup after the flash (IROM/DROM) is
+// initialized and the stack pointer has been set.
 //
 //export main
 func main() {
-	// Disable the protection on the watchdog timer (needed when started from
-	// the bootloader).
-	esp.RTC_CNTL.RTC_WDTWPROTECT.Set(0x050D83AA1)
+	// This initialization configures the following things:
+	// * It disables all watchdog timers. They might be useful at some point in
+	//   the future, but will need integration into the scheduler. For now,
+	//   they're all disabled.
+	// * It sets the CPU frequency to 160MHz, which is the maximum speed allowed
+	//   for this CPU. Lower frequencies might be possible in the future, but
+	//   running fast and sleeping quickly is often also a good strategy to save
+	//   power.
+	// TODO: protect certain memory regions, especially the area below the stack
+	// to protect against stack overflows. See
+	// esp_cpu_configure_region_protection in ESP-IDF.
 
-	// Disable both watchdog timers that are enabled by default on startup.
-	// Note that these watchdogs can be protected, but the ROM bootloader
-	// doesn't seem to protect them.
+	// Disable RTC watchdog.
+	esp.RTC_CNTL.RTC_WDTWPROTECT.Set(0x50D83AA1)
 	esp.RTC_CNTL.RTC_WDTCONFIG0.Set(0)
-	esp.TIMG0.WDTCONFIG0.Set(0)
+	esp.RTC_CNTL.RTC_WDTWPROTECT.Set(0x0) // Re-enable write protect
 
-	// Switch SoC clock source to PLL (instead of the default which is XTAL).
-	// This switches the CPU (and APB) clock from 40MHz to 80MHz.
-	// Options:
-	//   RTC_CNTL_CLK_CONF_SOC_CLK_SEL:       PLL    (default XTAL)
-	//   RTC_CNTL_CLK_CONF_CK8M_DIV_SEL:      2      (default)
-	//   RTC_CNTL_CLK_CONF_DIG_CLK8M_D256_EN: Enable (default)
-	//   RTC_CNTL_CLK_CONF_CK8M_DIV:          DIV256 (default)
-	// The only real change made here is modifying RTC_CNTL_CLK_CONF_SOC_CLK_SEL,
-	// but setting a fixed value produces smaller code.
+	// Disable Timer 0 watchdog.
+	esp.TIMG1.WDTWPROTECT.Set(0x50D83AA1) // write protect
+	esp.TIMG1.WDTCONFIG0.Set(0)           // disable TG0 WDT
+	esp.TIMG1.WDTWPROTECT.Set(0x0)        // Re-enable write protect
 
-	// TODO: Implement this for esp32s3
-	//esp.RTC_CNTL.RTC_CLK_CONF.Set((esp.RTC_CNTL_CLK_CONF_SOC_CLK_SEL_PLL << esp.RTC_CNTL_CLK_CONF_SOC_CLK_SEL_Pos) |
-	//	(2 << esp.RTC_CNTL_CLK_CONF_CK8M_DIV_SEL_Pos) |
-	//	(esp.RTC_CNTL_CLK_CONF_DIG_CLK8M_D256_EN_Enable << esp.RTC_CNTL_CLK_CONF_DIG_CLK8M_D256_EN_Pos) |
-	//	(esp.RTC_CNTL_CLK_CONF_CK8M_DIV_DIV256 << esp.RTC_CNTL_CLK_CONF_CK8M_DIV_Pos))
+	esp.TIMG0.WDTWPROTECT.Set(0x50D83AA1) // write protect
+	esp.TIMG0.WDTCONFIG0.Set(0)           // disable TG0 WDT
+	esp.TIMG0.WDTWPROTECT.Set(0x0)        // Re-enable write protect
 
-	// Switch CPU from 80MHz to 160MHz. This doesn't affect the APB clock,
-	// which is still running at 80MHz.
-	// TODO: Implement this for esp32s3
-	// esp.DPORT.CPU_PER_CONF.Set(esp.DPORT_CPU_PER_CONF_CPUPERIOD_SEL_SEL_160)
+	// Disable super watchdog.
+	esp.RTC_CNTL.RTC_SWD_WPROTECT.Set(0x8F1D312A)
+	esp.RTC_CNTL.RTC_SWD_CONF.Set(esp.RTC_CNTL_RTC_SWD_CONF_SWD_DISABLE)
+	esp.RTC_CNTL.RTC_SWD_WPROTECT.Set(0x0) // Re-enable write protect
 
-	// Clear .bss section. .data has already been loaded by the ROM bootloader.
-	// Do this after increasing the CPU clock to possibly make startup slightly
-	// faster.
+	// // Change CPU frequency from 20MHz to 80MHz, by switching from the XTAL to
+	// // the PLL clock source (see table "CPU Clock Frequency" in the reference
+	// // manual).
+	// esp.SYSTEM.SYSCLK_CONF.Set(1 << esp.SYSTEM_SYSCLK_CONF_SOC_CLK_SEL_Pos)
+
+	// // Change CPU frequency from 80MHz to 160MHz by setting SYSTEM_CPUPERIOD_SEL
+	// // to 1 (see table "CPU Clock Frequency" in the reference manual).
+	// // Note: we might not want to set SYSTEM_CPU_WAIT_MODE_FORCE_ON to save
+	// // power. It is set here to keep the default on reset.
+	// esp.SYSTEM.CPU_PER_CONF.Set(esp.SYSTEM_CPU_PER_CONF_CPU_WAIT_MODE_FORCE_ON | esp.SYSTEM_CPU_PER_CONF_PLL_FREQ_SEL | 1<<esp.SYSTEM_CPU_PER_CONF_CPUPERIOD_SEL_Pos)
+
 	clearbss()
-
-	// Initialize UART.
-	machine.InitSerial()
 
 	// Initialize main system timer used for time.Now.
 	initTimer()
@@ -153,3 +64,17 @@ func main() {
 	// Fallback: if main ever returns, hang the CPU.
 	exit(0)
 }
+
+func abort() {
+	// lock up forever
+	print("abort called\n")
+}
+
+//go:extern _vector_table
+var _vector_table [0]uintptr
+
+//go:extern _sbss
+var _sbss [0]byte
+
+//go:extern _ebss
+var _ebss [0]byte
